簡單整理一下剖析程式
