+++ 臺灣言語工具/剖析/中研院/剖析結構化工具.py
@@ -16,7 +16,6 @@
 
 感謝您的使用與推廣～～勞力！承蒙！
 """
-from 剖析相關工具.剖析工具 import 剖析工具
 import sys
 
 class 剖析結構化工具:
@@ -95,22 +94,25 @@ class 剖析結構化工具:
 		print(型體佮詞性語意[0], end = ' ', file = 目的)
 
 if __name__ == '__main__':
-	工具 = 剖析工具()
+# 	工具 = 剖析工具()
 # 	剖析結果字串集=工具.剖析('我想吃飯，，，我想吃很多飯。假如我也用這種方式旅行。再想到蝴蝶會生滿屋的毛蟲。')
-	剖析結果字串集 = ['#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#，(COMMACATEGORY)',
-			'#2:1.[0] %()#，(COMMACATEGORY)',
-			'#3:1.[0] %()#，(COMMACATEGORY)',
-			'#4:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
-			'#5:1.[0] S(C:假如|NP(Head:N:我)|ADV:也|PP(Head:P:用|NP(DM:這種|Head:N:方式))|Head:Vi:旅行)#。(PERIODCATEGORY)',
-			'#6:1.[0] VP(ADV:再|Head:Vt:想到|NP(S‧的(head:S(NP(Head:N:蝴蝶)|ADV:會|Head:Vt:生|NP(Head:N:滿屋))|Head:T:的)|Head:N:毛蟲))#。(PERIODCATEGORY)',
-			'#1:1.[0] VP(evaluation:Dbb:再|Head:VE2:想到|goal:NP(predication:S‧的(head:S(agent:NP(Head:Nab:蝴蝶)|epistemics:Dbaa:會|Head:VC31:生|theme:NP(Head:Na:滿屋))|Head:DE:的)|Head:Nab:毛蟲))#。(PERIODCATEGORY)']
+	剖析結果字串集 = [
+			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#',
+ 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
+ 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
+			
+	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
+	 			
+			'#1:1.[0] %(NP(Head:N:我)|Vt:想|COLONCATEGORY::)#&gt;(PARENTHESISCATEGORY)',
+			
+]
 #  	print(剖析結果字串集)
 	結構化工具 = 剖析結構化工具()
 # 	print(國閩單位翻譯(('吃',)))
 	for 剖析結果字串 in 剖析結果字串集:
 		結構化結果 = 結構化工具.結構化剖析結果(剖析結果字串)
-# 		print(結構化結果)
-		結構化工具.處理結構化結果(結構化結果, 結構化工具.印出)
+		print(結構化結果)
+# 		結構化工具.處理結構化結果(結構化結果, 結構化工具.印出)
 		print()
 # 		翻譯結果 = 結構化工具.處理結構化結果(結構化結果, 國閩單位結構化翻譯)
 # # 		print(翻譯結果)
